add st7735 driver

+++ examples/ssd1306/ssd1306.ino
@@ -14,7 +14,7 @@ void setup() {
 }
 
 void loop() {
-    screen.display(0, 0, 128, 64, 0);
+    screen.display(0, 0, 128, 64, (uint8_t)0);
 
     delay(1000);
